attack-map-main/requirements.txt
streamlit
pydeck
pandas
altair
requests
boto3
streamlit-autorefresh
python-dotenv
schedule
maxminddb
pycountry
scikit-learn